+++ pyproject.toml
@@ -33,7 +33,7 @@ dependencies = [
 [project.scripts]
 flyte = "flyte.cli.main:main"
 a0 = "flyte._bin.runtime:main"
-clustered = "flyte._bin.runtime:clustered_main"
+clustered = "flyte._bin.clustered:clustered_main"
 fserve = "flyte._bin.serve:main"
 c0 = "flyte._bin.connect:main"
 flyte-mcp = "flyte._bin.mcp:main"
